refactor: extract shared LLM utilities to llm_utils.py
Move get_startwith, get_model_key_from_env, get_api_key_dict_from_env,
get_config_class_dict, and get_config_class from entrypoint.py into a
new llm_utils.py module.

This eliminates the circular import between entrypoint.py and
prompt_pipeline/entrypoint.py — the latter no longer needs to import
from the tutor's entrypoint via sys.path hacks.  Both consumers now
import from llm_utils, which has no dependency on either entrypoint.

Also fixes prompt_pipeline/entrypoint.py sys.path resolution: falls
back to the repo root when the Docker-only ai_tutor/ directory is
absent (fixes ModuleNotFoundError in CI).

All 231 tests pass.
